+++ R/fftfilt.R
@@ -20,6 +20,8 @@
 # 20200417  GvB       setup for gsignal v0.1.0
 # 20200420  GvB       adapted slightly (rounding in case of whole numbers)
 # 20201121  GvB       done away with FFTfilt, only method for Ma()
+# 20210630  GvB       fixed Github bug #3: Problems with fftfilt when FFT
+#                     length is provided by user
 #------------------------------------------------------------------------------
 
 #' FFT-based FIR filtering
@@ -141,9 +143,10 @@ fftfilt.default <- function(b, x, n = NULL) {
     if (is.vector(x)) {
       R <- ceiling(lx / L)
       y <- rep(0L, lx)
+      B <- stats::fft(postpad(b, n))
       for (r in seq_len(R)) {
         lo <- (r - 1) * L + 1
-        hi <- min(r * L, nrx)
+        hi <- min(r * L, lx)
         tmp <- rep(0L, n)
         tmp[1:(hi - lo + 1)] <- x[lo:hi]
         tmp <- ifft(stats::fft(postpad(tmp, n)) * B)
@@ -153,6 +156,7 @@ fftfilt.default <- function(b, x, n = NULL) {
     } else {
       R <- ceiling(nrx / L)
       y <- matrix(0L, nrx, ncx)
+      B <- stats::fft(postpad(b, n))
       for (r in seq_len(R)) {
         lo <- (r - 1) * L + 1
         hi <- min(r * L, nrx)